+++ oop/student-management/src/main/java/com/example/studentmanagement/Model/StudentResponse.java
@@ -20,7 +20,7 @@ public class StudentResponse {
     public StudentResponse getStudentResponseByType(String type) {
         switch (type) {
             case Utilities.DEFAULT: {
-                return new StudentResponse(false, new Student(-1,"","","",false,-1,-1, (long) -1,"","",""));
+                return new StudentResponse(false, new Student((long) -1,"","","",false,-1,-1, (long) -1,"","",""));
             }
             default: {
                 return new StudentResponse();
